2nd code cleanup of XPU profiler for incoming scope profiler (#1256)
Summary:
**Summary:**

2nd code cleanup of XPU profiler plugin to prepare it for incoming PR with implementation of XPU scope profiler.

**Changes:**

- separate profiler and session classes
- simplify macro for error handling

Pull Request resolved: https://github.com/pytorch/kineto/pull/1256

Reviewed By: ryanzhang22

Differential Revision: D93888612

Pulled By: divyanshk

fbshipit-source-id: 4381d140d24b05836c79f1aad4a5e6f4d9172265
(cherry picked from commit e8d2d27d841dc7d3f16bba583f47e9ef9b6b7c08)
