+++ libkineto/src/plugin/xpupti/XpuptiProfilerMacros.h
@@ -8,10 +8,9 @@
 
 #pragma once
 
-#include <stdexcept>
-#include <string>
+#include <string_view>
 
-#include <pti/pti_version.h>
+#include <pti/pti.h>
 
 namespace KINETO_NAMESPACE {
 
@@ -21,36 +20,20 @@ using namespace libkineto;
   (PTI_VERSION_MAJOR > MAJOR ||            \
    (PTI_VERSION_MAJOR == MAJOR && PTI_VERSION_MINOR >= MINOR))
 
-#if PTI_VERSION_AT_LEAST(0, 10)
-#define XPUPTI_CALL(returnCode)                                               \
-  {                                                                           \
-    if (returnCode != PTI_SUCCESS) {                                          \
-      std::string funcMsg(__func__);                                          \
-      std::string line(std::string(" line ") + std::to_string(__LINE__));     \
-      std::string codeMsg = std::to_string(returnCode);                       \
-      std::string HeadMsg("Kineto Profiler on XPU got error from function "); \
-      std::string Msg(". The error code is ");                                \
-      std::string detailMsg(". The detailed error message is ");              \
-      detailMsg = detailMsg + std::string(ptiResultTypeToString(returnCode)); \
-      throw std::runtime_error(                                               \
-          HeadMsg + funcMsg + line + Msg + codeMsg + detailMsg);              \
-    }                                                                         \
-  }
-#else
-#define XPUPTI_CALL(returnCode)                                               \
-  {                                                                           \
-    if (returnCode != PTI_SUCCESS) {                                          \
-      std::string funcMsg(__func__);                                          \
-      std::string line(std::string(" line ") + std::to_string(__LINE__));     \
-      std::string codeMsg = std::to_string(returnCode);                       \
-      std::string HeadMsg("Kineto Profiler on XPU got error from function "); \
-      std::string Msg(". The error code is ");                                \
-      throw std::runtime_error(HeadMsg + funcMsg + line + Msg + codeMsg);     \
-    }                                                                         \
-  }
-#endif
+[[noreturn]] void throwXpuRuntimeError(
+    std::string_view errMsg,
+    pti_result errCode);
 
-class XpuptiActivityApi;
+[[noreturn]] void
+throwXpuRuntimeError(const char* func, int line, pti_result errCode);
+
+#define XPUPTI_CALL(returnCode)                             \
+  {                                                         \
+    if (returnCode != PTI_SUCCESS) {                        \
+      throwXpuRuntimeError(__func__, __LINE__, returnCode); \
+    }                                                       \
+  }
+
 using DeviceIndex_t = int8_t;
 
 } // namespace KINETO_NAMESPACE
